--- src/main/java/focusedCrawler/crawler/async/cookieHandler/CookieHandler.java
@@ -1,7 +1,0 @@
-package focusedCrawler.crawler.async.cookieHandler;
-
-// Common Interface for cookie handlers for both fetchers
-
-public interface CookieHandler {
-
-}
