+++ src/main/java/focusedCrawler/crawler/crawlercommons/fetcher/http/SimpleHttpFetcher.java
@@ -93,7 +93,7 @@ import org.apache.tika.metadata.Metadata;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import focusedCrawler.crawler.async.cookieHandler.ConcurrentCookieJar;
+import focusedCrawler.crawler.async.cookies.ConcurrentCookieJar;
 import focusedCrawler.crawler.crawlercommons.fetcher.AbortedFetchException;
 import focusedCrawler.crawler.crawlercommons.fetcher.AbortedFetchReason;
 import focusedCrawler.crawler.crawlercommons.fetcher.BadProtocolFetchException;
